New Order, Ticket Repositiories and Controllers added, added endpoint for getting available seats in showing

+++ backend/src/main/java/com/csci4050/promotion/entities/Promotion.java
@@ -23,6 +23,6 @@ public class Promotion {
     private Integer id;
 
     private String event;
-    private float discount;
+    private double discount;
     private boolean active;
 }
